+++ src/main/java/com/ecommerce/motomart/Services/UserService.java
@@ -1,8 +1,10 @@
 package com.ecommerce.motomart.Services;
 
 import com.ecommerce.motomart.DTO.UserDTO;
+import com.ecommerce.motomart.Exceptions.UserNotFoundException;
 import com.ecommerce.motomart.Models.User;
 import com.ecommerce.motomart.Repositories.UserRepository;
+import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,6 +17,9 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
+    @Autowired
+    private ModelMapper modelMapper;
+
     public List<UserDTO> getAllUsers() {
         return userRepository.findAll().stream()
                 .map(this::convertToDTO)
@@ -22,43 +27,42 @@ public class UserService {
     }
 
     public UserDTO getUserById(Long id) {
-        return convertToDTO(userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found")));
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException(id));
+        return convertToDTO(user);
     }
 
     public UserDTO createUser(UserDTO userDTO) {
         User user = convertToEntity(userDTO);
-        return convertToDTO(userRepository.save(user));
+        User createdUser = userRepository.save(user);
+        return convertToDTO(createdUser);
     }
 
     public UserDTO updateUser(Long id, UserDTO userDTO) {
-        User user = userRepository.findById(id).orElseThrow(() -> new RuntimeException("User not found"));
-        user.setUsername(userDTO.getUsername());
-        user.setEmail(userDTO.getEmail());
-        user.setPassword(userDTO.getPassword());
-        user.setRole(userDTO.getRole());
-        return convertToDTO(userRepository.save(user));
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException(id));
+
+        existingUser.setUsername(userDTO.getUsername());
+        existingUser.setEmail(userDTO.getEmail());
+        existingUser.setPassword(userDTO.getPassword());
+        existingUser.setRole(userDTO.getRole());
+
+        User updatedUser = userRepository.save(existingUser);
+        return convertToDTO(updatedUser);
     }
 
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
 
     private UserDTO convertToDTO(User user) {
-        UserDTO userDTO = new UserDTO();
-        userDTO.setUserId(user.getUserId());
-        userDTO.setUsername(user.getUsername());
-        userDTO.setEmail(user.getEmail());
-        userDTO.setPassword(user.getPassword());
-        userDTO.setRole(user.getRole());
-        return userDTO;
+        return modelMapper.map(user, UserDTO.class);
     }
 
     private User convertToEntity(UserDTO userDTO) {
-        User user = new User();
-        user.setUsername(userDTO.getUsername());
-        user.setEmail(userDTO.getEmail());
-        user.setPassword(userDTO.getPassword());
-        user.setRole(userDTO.getRole());
-        return user;
+        return modelMapper.map(userDTO, User.class);
     }
 }
